Various warning, logging, jit fixes

+++ Core/FileSystems/ISOFileSystem.cpp
@@ -414,7 +414,7 @@ size_t ISOFileSystem::ReadFile(u32 handle, u8 *pointer, s64 size)
 			posInSector = 0;
 			secNum++;
 		}
-		e.seekPos += size;
+		e.seekPos += (unsigned int)size;
 		return totalRead;
 	}
 	else
@@ -449,7 +449,7 @@ size_t ISOFileSystem::SeekFile(u32 handle, s32 position, FileMove type)
 			if (e.isRawSector)
 				e.seekPos = e.openSize;
 			else
-				e.seekPos = e.file->size + position;		
+				e.seekPos = (unsigned int)(e.file->size + position);
 			break;
 		}
 		return (size_t)e.seekPos;
